build-image: Add support for creating a rt image
This commit adds support for the creation of an image
with the rt kernel and modules. There is an independent
STX package list is added to support this.
'build-image' refers and reuses the base LAT yaml templates:
base-bullseye.yaml, base-initramfs-bullseye.yaml to create
the LAT yaml files for the rt image. All the std kernel names
will be replaced with rt kernel names. A complete package
list of rt image from 'stx-rt.lst' will be filled into the
field of 'packages:' of LAT yaml for rt.

Test Plan:
Pass: build-image --std
Pass: build-image --rt

Story: 2008846
Task: 44867

Depends-On: https://review.opendev.org/c/starlingx/tools/+/838625

Signed-off-by: hbai <haiqing.bai@windriver.com>
Change-Id: I9cf8c522834c8fdab760890d68b83e84797c35bb