Update build-helm-charts.sh to support FluxCD apps
The optional applications that are not delivered with the platform use
the build-helm-charts.sh to build their tarballs.

This updates the build script to support FluxCD apps and position the
script to easily remove the Armada related code when no longer needed.

The package provided to the script should install the FluxCD manifests
in './usr/lib/fluxcd'. The contents of that directory will be copied to
the top level 'fluxcd-manifests' directory within the tarball.

Followup commits will be provided address the following for FluxCD:
 - Merging manifests support via parse_yaml()
 - Image versions from record files via
   build_image_versions_to_manifest()
 - Support for Debian .deb packages

Test Plan:
PASS: Successfully build a test application and verify the resulting
      tarball contains the correct formatting
PASS: Run application-upload on the resulting tarball to verify
      formatting with the framework.

Task: 45114
Story: 2009138
Change-Id: I5405dee22cbb574b3683338d1f7d39018a95c595
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>
Signed-off-by: Robert Church <robert.church@windriver.com>