build-docker-images: allow docker cache
Add an option to docker image build scripts to enable docker cache. This
allows us to build images much more quickly when troubleshooting build
scripts.

Story: 2009897
Task: 45121

TESTS
========

Build base image and one of the docker apps with and without the
"--cache" option and make sure it works as expected

Signed-off-by: Davlet Panech <davlet.panech@windriver.com>
Change-Id: I0a1e94ebfe8a263c6ddd293d0da40b5eb8550e80