build-iso: Create package_checksums file within iso
Extended build-iso to create the package_checksums file
after the Packages subdirectory has been created and all rpms
are signed.  It can be found in the root directory of the
mounted iso.

The kickstarts will have to be extended to copy the
package_checksums from the iso to the final location
on the controller.

Story: 2009948
Task: 45219
Signed-off-by: Scott Little <scott.little@windriver.com>
Change-Id: I4cf064911600f89e356527ee274f40c6ce14a990