build-helm-charts: python 2/3 compatibility
Make sure the helm charts script works with both python2 and python3.
The referenced python scripts work with either python version, but some
of the required module versions are not available on Debian for python2.

Solution: look for python2 (for CentOS compatibility), fallback to python3.

TESTS
===========================================
Build helm charts on both CentOS and Debian

Story: 2009897
Task: 45411

Change-Id: I4df12f1654c0f6950d71b83188a454fc54449d71
Signed-off-by: Davlet Panech <davlet.panech@windriver.com>