build-tools: discover firmware files for installer
The script build-tools/update-pxe-network-installer allows one to add
additional firmware files to the installer by specifying an environment
variable, UPDATE_FW_LIST. This is inconvenient, because it requires the
user to know the exact path to the firmware file to be included.

Solution: search for files names "centos_firmware.inc" in the source
tree and take firmware files from there.

Story: 2010067
Task: 45529

Change-Id: I7673bd2c1bab4061936d8558a64f016a3661feec
Signed-off-by: Davlet Panech <davlet.panech@windriver.com>