debian: build-tools: skip "do-not-build" dirs
Ignore projects under directories named "do-not-build". This feature is
present in CentOS tools, but was missing in Debian tools. It allows one
to create repo manifests that include gits which should be skipped
during the build.

TESTS
==============================================

Build all packages and make sure the total number of packages built is
the same as before the patch.

Story: 2010055
Task: 45884

Signed-off-by: Davlet Panech <davlet.panech@windriver.com>
Change-Id: I396af5f1dcd09029681fcea193b02c89edc62e7d