debrepack.py: add checksum upon download
We use the return value of "wget" to check if download succeeds. However
we observed "wget" returns 0 with a 0-length .tar.gz file. It looks like
this is caused by "wget" when a connection to the server breaks in the
middle of download, and yet "wget" reports success (ie returns 0
status).

Adding the checksum upon download to check if the download file is
correct.

Test Plan:
Pass: downloader -s

Closes-bug: 1986862
Signed-off-by: Yue Tao <Yue.Tao@windriver.com>
Change-Id: I5a64f695c5e60ef306eed309ee923556aaa0fd0a