debian: build-tools: build stuck in chroot creation
Build-pkgs only catches the successful messages of chroot
creating in the chroot.log and returns, it does not catch
the failed or error messages in this log file. This will
cause the build-pkgs to get stuck and not to return from
waiting to show this log file if fails to creat chroot.
To fix this we make build-pkgs to quit after catching any
error message with 'E: ' prefix in chroot.log.

Test Plan:
Pass: Created a local repository and removed 'libc-dev-bin'
      from that without repository updating. Set this
      repository to chroot creating.  Remove all the old
      chroots, then run 'build-pkgs -c -p <package name>'

Closes-bug: 1986864
Signed-off-by: Wentao Zhang <Wentao.Zhagn@windriver.com>
Change-Id: Ia0ed99a9ccfbe96b169c5ebf4887745d6321bb2d