build-pkgs: Fixed the issue that returned chroot is None
If pkgbuilder fails to execute the builder's commands
and fails to assign idle chroot for package build,
not only the 'status' in the returned value should be
set to 'fail', but also the 'chroot' should be used to
set to simple words to show the error type.
And the code is improved to use the 'chroot' safely.

Test Plan:
Only run 'build-pkgs -a' multiple times for this issue
hard to reproduce.

Closes-bug: 1987906
Signed-off-by: Haiqing Bai <haiqing.bai@windriver.com>
Change-Id: Ia53860a983119ddedac870c39c840e2e1377b46d