debrepack: checksum() and set_revision() covers stx/git/<pkg>
Following CentOS, introudce GITREVCOUNT, SRC_DIR and BASE_SRCREV in
meta_data.yaml to count the git repo in set_revision(), so that the
change in git repo is reflected in the package version.

If a package with local git repo, and the git repo is updated but no
change in debian folder, the checksum of the package is no changed,
which causes the "build-pkgs -a" or "build-pkgs -p xx" doesn't rebuild
the package.

Improve the checksum() to cover the change in local git repo, that adds
the last 10 commits ("git log --oneline -10") and local change ("git
diff") to checksum.

Test Plan:

Pass: get different package version and checksum

Add the changes to ceph/ceph/debian/meta_data.yaml
=================================================
 revision:
   dist: $STX_DIST
   PKG_GITREVCOUNT: true
+  GITREVCOUNT:
+    SRC_DIR: ${STX_GIT_SRC_DIR}/ceph
+    BASE_SRCREV: ca74598065096e6fcbd8433c8779a2be0c889351
=================================================

$build-pkgs -p ceph

Story: 2008846
Task: 46198

Signed-off-by: Yue Tao <yue.tao@windriver.com>
Change-Id: I5c5756423c6ea210942f00b6b3300c11ee20a859