Debian: introduce debian-image.inc as the installing list
"base-bullseye.lst" is both the download list and installing list. It
can't be implemented if someone wants to download a binary but not
install it to image. According to:

https://docs.starlingx.io/developer_resources/Layered_Build.html#controlling-the-package-content-of-the-iso

In CentOS version, we use the "image.inc" file as the installing list.
Other .lst files are the download lists. Also introducing such file
"debian-image.inc" in Debian.

The purpose of this commit is to separate the 2 lists ASAP, so that we
can add build dependencies into "base-bullseye.lst" right now, that is
requested to fix CVEs. So this commit just copies the
"base-bullseye.lst" to "debian-image.inc", and removes the versions.

A subsequence commit will remove the dependencies from
"debian-image.inc".

Test Plan:

Pass: build-image
Pass: comparing the installing packages after and before
      (No difference of "dpkg-query -l")

Story: 2008846
Task: 46274

Signed-off-by: Yue Tao <yue.tao@windriver.com>
Change-Id: Ic7af95cbefc80dff65b255028c78dc0cd078ac08