Add zerorpc wheel package
Build of docker image stx-keystone-api-proxy fails on CentOS, since
the docker image includes sysinv, which requires zerorpc.

The code for sysinv may still import zerorpc, but the image will not
run sysinv or zerorpc, so this fix is sufficient to get the image to
build.

TEST PLAN:
PASS: Build stx-keystone-api-proxy docker image on CentOS

Closes-Bug: 1997977

Signed-off-by: Alyson Deives Pereira <alyson.deivespereira@windriver.com>
Change-Id: I30da6ec139f4cd31759eb249f008e1101e02e713