Debian pre-patch-iso ostree sha update
This commit updates the ostree_repo inside the
iso to execute a pull-local to update the iso
ostree_repo with the patch commit, this is required
because whenever we reuse the initramfs the top
commit on the build ostree_repo changes so the
pull-local fixes it. If we have a new initramfs the
pull-local won't do any changes.

Test Plan:
Pass: Patch an iso and install it
Pass: check sw-patch query output
Pass: check ostree log starlingx history
Pass: compare build ostree with iso ostree

Story: 2009969
Task: 46467
Signed-off-by: Luis Sampaio <luis.sampaio@windriver.com>
Change-Id: I1e992ce2e3ba54013a043b776981e7994d7fc9db