docker-images: Add support for LOCI's UID/GID args
In [1], a patch that adds the user/group `keystone` with UID/GID 42424
was introduced to be applied during the Debian base image build.

As a result, every LOCI-based image built on top of it will have
a UID/GID conflict with the user in question, since this same value is
the default used by LOCI when creating users ([2]).

Since LOCI already allows the customization of these build arguments,
this change adds corresponding support for them in the build script.

In order to avoid conflict with the existing Bash readonly variable UID,
the arguments must be expressed as `PROJECT_UID` and `PROJECT_GID` in
the container build recipes, so it can be safely imported into the build
script.

[1] https://review.opendev.org/c/starlingx/integ/+/854246
[2] https://github.com/openstack/loci

Test Plan:
PASS: Without using the build arguments `PROJECT_UID` and `PROJECT_GID`,
      verify that LOCI is defaulting to 42424 during the build.
PASS: Using the build arguments `PROJECT_UID` and `PROJECT_GID`, verify
      that LOCI is using different values during the build.

Signed-off-by: Luan Nunes Utimura <LuanNunes.Utimura@windriver.com>
Change-Id: Ibb4b0cbcdb0f3e9440810339408fc96ab356f55b