Update mysql-client in bindep for dpkg placement
While building one of the stx-openstack images - stx-placement - it was
observed that the current bindep file for LOCI carries an outdated
version of the mysql-client package.

According to [1], the mysql-client package currently only exists for
Debian stretch and unstable (sid). However, by taking a look at its
specification, this package is actually a mirror for another package,
default-mysql-client, which exists for Debian stretch, buster, bullseye,
bookwork and sid.

Since this package is only used for the stx-placement build, we can
safely replace it and ensure that this image is buildable not only on
Debian stretch but also on other versions as well.

[1] https://packages.debian.org/search?suite=all&arch=any&mode=path&searchon=names&keywords=mysql-client

Test Plan:
PASS: Without this change, verify that stx-placement does not build due
      to: "Package mysql-client has no installation candidate"
PASS: With this change, verify that stx-placement was successfully built
      and that the bindep file was updated accordingly

Story: 2010072
Task: 47089

Signed-off-by: Luan Nunes Utimura <LuanNunes.Utimura@windriver.com>
Change-Id: I0292ac399262ac3e55bdb0f4b0093c8ce689a993