debrepack.py ignore .git on checksum calculation
This commit updates the debrepack checksum method
to ignore .git files in the calculation. This issue happens
for packages that have a git directory in the src_path
meta_data.yaml  and even tough we have the same files
the checksum won't match if we are comparing 2 branches
because of the .git files (.gitreview for example won't be the
same).

Test Plan:
Pass: build-pkgs
Pass: build-pkgs --reuse

Story: 2008846
Task: 47236
Signed-off-by: Luis Sampaio <luis.sampaio@windriver.com>
Change-Id: I9c5d2f4d268def30abfbd1174e0389d523568c93