Update to use srcdir when calculating checksum
Update the debrepack checksum function to use
the src_dir instead of the whole git when calculating
the checksum of a package. This will allow us to use
GITREVCOUNT + SRC_DIR in the meta_data.yaml and
changes to other directories in the git should  not
trigger a checksum missmatch.

Story: 2010550
Task: 47306
Signed-off-by: Luis Sampaio <luis.sampaio@windriver.com>
Change-Id: I978ba35133979e7b2819157ae9be91ce8b0f5315