Change version of netaddr
Netaddr is using version 0.7.19 on remote cli. Because of this
version, commands being executed towards platform apis are causing a
syntax warning.
This fix changes the minimum requirement of netaddr to 0.7.20 to
address a fix for this warning.

This file is referenced by the build-wheels script which uses the
version here to build the wheel tarball that is used by the loci
builder when building platformclients.

Test Plan:
PASS: Build debian iso and perform fresh install.
PASS Build python3 wheels tarball and build
stx-platformclients image on Debian.

Partial-Bug: 1999563

Signed-off-by: Luiz Felipe Kina <LuizFelipe.EiskeKina@windriver.com>
Change-Id: I058f33e08d03dbd309c908060ccdcf2cb002d009