debrepack.py ignore pyc on checksum calculation
This commit updates the debrepack checksum method
to ignore pyc files in the calculation to avoid checksum
missmatch if we have __pycache__/*.pyc or *.pyo files in
the repo after building a patch.

The commit also updates the checksum calc when using
GITREVCOUNT to set the SHA char to 10 to avoid local
git config displaying different number of characters in
the calculation.

Story: 2008846
Task: 47470
Signed-off-by: Luis Sampaio <luis.sampaio@windriver.com>
Change-Id: I1d21a52b192fb1517471b307f7991c07952531bb