build-image: change owner for LAT deploy directory
The LAT deploy directory's owner is root because lat container's
default user is root.
The old version LAT deploy directory's attribute is rwxrwxrwx
because when secure boot is enabled, LAT changes the attribute
from rwxr-xr-x to rwxrwxrwx before writing sig files of initramfs
to the deploy dir.
In the secure boot enhancement commits, the default secure
boot switch setting EFI_SECURE_BOOT is changed to "disable".
We also remove all the hard coded signings in LAT including
the attribute changes because they aren't needed any more.
The signings are done in the LAT hook functions now.

But the signings of iso (especially the one done in jenkins
script) need write the sig file into deploy dir. We decide not
to touch LAT again and change the owner of deploy directory
to the default user of builder in build-image.

Test Plan:
Pass: build-image
Pass: check the user/group of /localdisk/deploy and the
      files/directories under it are right.

Story: 2008846
Task: 46366

Signed-off-by: Li Zhou <li.zhou@windriver.com>
Change-Id: I6de16e43447116a7bbff08d820336d147725d628