Upversion os-vif to 2.0.1 for PY3.8/3.9
After migrating stx-openstack images from CentOS to Debian, it was
observed that, on Debian, the `os-vif` library* is currently one patch
version behind the one on CentOS.

The reason behind this is the fact that the upper constraints file is
currently fixing the library version at 2.0.0 for PY3.8/PY3.9, while for
PY3.6/3.7, the version is fixed at 2.0.1.

Since this patch fixes the library compatibility with certain NICs after
the Linux kernel 5.8 release ([1] and [2]), and was already present in
the CentOS-based stx-openstack images, this change proposes the
upversion of the library for the Debian-based images as well.

* `os-vif` is an integration library between network and compute
  providers, i.e., stx-neutron and stx-nova.

[1] https://docs.openstack.org/releasenotes/os-vif/ussuri.html#relnotes-2-0-1-stable-ussuri
[2] https://bugs.launchpad.net/os-vif/+bug/1892132

Test Plan (on AIO-SX):
PASS - Build stx-neutron image
PASS - Build stx-nova image
PASS - Upload built images to the local registry, perform helm overrides
       to change the container images and reapply stx-openstack
PASS - Verify that affected pods are up and running properly
PASS - Verify that the `os-vif` library was successfully upgraded by
       running `pip freeze | grep os-vif` in one of the affected
       containers

Story: 2010317
Task: 47623

Signed-off-by: Luan Nunes Utimura <LuanNunes.Utimura@windriver.com>
Change-Id: Ic15759a72fc629300de95e90995ce1b86b0c7152