Update static tag for snmp images
In order to fix the CVEs reported on snmp container images, an
update on the tags needs to be done.
The 20230412T060000Z.0 build contains updated packages that fix
a big number on CVEs reported on stx.8.0-v1.0.0 tag.

The tags will be updated to stx.9.0-v1.0.1 for stx-snmp and
stx-fm-trap-subagent images.
(Note: stx-fm-subagent was already updated in
https://review.opendev.org/c/starlingx/root/+/880335)

Story: 2010710
Task: 47871

Signed-off-by: Enzo Candotti <enzo.candotti@windriver.com>
Change-Id: I34395cab4a19fc21a1dc5b2479f9fccc83c56b04