build-helm-charts: Fix --package option parsing
When building helm charts, it has been observed that the preferred long
option for specifying packages, `--package`, isn't being recognized by
`getopt` as a valid option.

As a result, one must resort to the deprecated alternative `--rpm`.

Therefore, this change adds `package` as one of the long options for
`getopt`. And since it updates the parameters of this build script, it
also updates the help message of the `--os` parameter to match the other
scripts:

  * build-wheels/build-wheel-tarball.sh;
  * build-docker-images/build-stx-base.sh;
  * build-docker-images/build-stx-images.

Test Plan:
PASS - Build helm charts with --rpm
PASS - Build helm charts with --package

Closes-Bug: 2022090

Change-Id: Ia45930d30e99ceb15d2f0753519cd8290fcaa446
Signed-off-by: Luan Nunes Utimura <LuanNunes.Utimura@windriver.com>