Update patch xml schema to use id and sw_version
This commit removes the patch_version tag which will not be
required by the patch framework and renames the patch_id
tag to just 'id'.

The patch version will be used in the sw_version tag, e.g:
- 1.0.0 from release version
- 1.0.1 for the 1st patch for this release

The patch id will be generated based on component and
sw_version. The id is also the file name:
$id.patch

Test Plan:
- Enter the builder container
$ stx shell
$ cd $MY_REPO/build-tools/stx/patch

- Build patch
$ ./patch_builder.py --recipe patch-recipe.xml
- Verify metadata has the updated schema

Story: 2010676
Task: 48640
Change-Id: I311438403560f04d6d9dd32778af5c6da101b7ac
Signed-off-by: lsampaio <luis.sampaio@windriver.com>