Build neutron-lib wheels from 3.4.0 to 3.4.2
The master branch build of the stx-neutron container is failing on a
dependency python issue. The neutron 22.1.1.dev85 version modified the
minimum required version for the neutron-lib package, and that is
causing the build of the stx-neutron image to fail.

This change updates the wheels of the neutron-lib package to version
3.4.2, in order to fix the build for this image.

Test plan:
PASS: Build wheels
PASS: Build stx-neutron image
PASS: Build all images

Closes-Bug: 2052644

Change-Id: I9ef49300898b6697852e9f3aa6163b6b684d2803
Signed-off-by: Lucas de Ataides <lucas.deataidesbarreto@windriver.com>